refactor: Jackson Kotlin module and refactor ShortenRequest class
Added the "jackson-module-kotlin" dependency to support JSON parsing in Kotlin. It is essential to parse JSON data and convert it into Kotlin data classes seamlessly.

Additionally, the unnecessary private constructor in the ShortenRequest class was removed to simplify the code. This constructor was superfluous and potentially misleading as it suggests that the class can be instantiated without providing a longUrl, which is not the case.